added dynamic paths, it is necessary to define in .bashrc the export e.g.
- added some flag support for gcc-9

export casadi_path=/home/rslstudent/Students/Emanuel/casadi-3.6.6-linux64-matlab2018b

+++ main_c/main_mpc_udp.c
@@ -131,7 +131,7 @@ void *update_data(void *arg) {
         fflush(stdout);
         
         if (flag != 0) {
-            printf("Error in MPC8: flag = %d\n", flag);
+            printf("Error in MPC8: flag = %d\n", (int) flag);
             shared.should_exit = 1;
             pthread_mutex_unlock(&shared.mutex);
             return NULL;
